Address code review suggestions: fix imports, complete protocol, move etcd config test
- Move timedelta and HTTPStatusError to top-level imports in http_testing.py
- Complete HTTPResponseProtocol: add status_code, content, text, headers,
  json(), raise_for_status(), close() — previously the protocol only declared
  streaming methods and had no practical type safety
- Remove inaccurate comment from test_authtoken.py (claimed MockHTTPResponse
  was a "drop-in replacement" and "httpx compatible")
- Move etcd test_config out of test_integration.py into a new test_unit.py
  so it no longer requires a docker environment to run
- Add mock_http fixture test (verifies PropertyMock patch reaches check.http)
- Add MockHTTPResponse tests: iter_content chunking, iter_lines decode_unicode
  and custom delimiter, file_path constructor, raise_for_status 2xx no-op
